% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_from_example_to_function.R
\name{from_example_to_function}
\alias{from_example_to_function}
\title{From example to function}
\usage{
from_example_to_function(expr, env = parent.frame())
}
\arguments{
\item{expr}{The expression user wants to transform to a function}

\item{env}{Working environment}
}
\value{
Function 1L. Output function.
}
\description{
From a user exemple code (expr), get the corresponding function
}
