% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_from_function_to_mask.R
\name{from_function_to_mask}
\alias{from_function_to_mask}
\title{From function to mask}
\usage{
from_function_to_mask(func)
}
\arguments{
\item{func}{Function 1L. Input function}
}
\value{
data.frame. Corresponding mask
}
\description{
Transform a function into a genproc compatible mask
}
