% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_demo_framework.R
\name{create_demo_framework}
\alias{create_demo_framework}
\title{create_demo_framework}
\usage{
create_demo_framework(demo_path)
}
\arguments{
\item{demo_path}{Folder path in which you want to create the demo elements. Those will be encapsulated in an unique folder (in demo_path)}
}
\value{
Create, within the designated path: R datasets, mask, function and args mapping
}
\description{
Given a path, create the necessary framework for PoC demo
 (file conversion)
}
